backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
alembic
email-validator
python-jose[cryptography]
python-multipart
passlib
openpyxl
reportlab
psycopg[binary]